extract `setupRepo` and `humanOp` from experiment.spec.ts
